BackendPython/Marketplace/requirements.txt
asgiref==3.8.1
attrs==25.1.0
boto3==1.37.25
botocore==1.37.25
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==4.2.19
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
dnspython
drf-yasg
exceptiongroup
h11
idna
imports
inflection==0.5.1
jmespath==1.0.1
mysqlclient==2.2.7
outcome==1.3.0.post0
packaging==24.2
pillow
PyJWT==2.9.0
pymongo
PySocks
python-dateutil
python-decouple
python-dotenv
pytz
PyYAML
requests
s3transfer
selenium
six
sortedcontainers
sqlparse
stripe
trio
trio-websocket
typing_extensions
uritemplate
urllib3
websocket-client
wsproto
imageio
imageio-ffmpeg
django-storages
opencv-python
moviepy
validators
geopy
daphne
channels
channels_redis
babel
django-celery-beat
